PythonClassifierService/classifiers/requirements.txt
asgiref==3.5.2
backports.zoneinfo==0.2.1
certifi==2022.6.15
charset-normalizer==2.1.0
Django==4.0.6
djangorestframework==3.13.1
idna==3.3
numpy==1.23.0
opencv-python==4.6.0.66
Pillow==9.2.0
pytz==2022.1
requests==2.28.1
sqlparse==0.4.2
torch==1.12.0
torchvision==0.13.0
typing_extensions==4.3.0
urllib3==1.26.10
